SRVCOM-1511 Use topology list URL conditionally
The topology URL has change and from 4.9+ old URL isn't longer working.
New URL needs to be used.

Also, to do it a code refactor was done. Now, it's much easier to
conditionally write code without Cypress promises which are hard to
chain and not required to resolve OCP version compatibility. Also, it's
easier to navigate the code in IDE, because of regular classes.
