+++ test/ui/cypress/integration/serving.spec.js
@@ -1,140 +1,32 @@
+import Environment from "../code/environment";
+import ShowcaseKservice from "../code/knative/serving/showcase";
+import OpenshiftConsole from "../code/openshift/openshiftConsole";
+
 describe('OCP UI for Serverless', () => {
 
-  class ShowcaseKservice {
-    constructor(ops = {}) {
-      this.app = ops.app || 'demoapp'
-      this.name = ops.name || 'showcase'
-      this.namespace = ops.namespace || Cypress.env('TEST_NAMESPACE')
-      this.image = ops.image || {
-        // TODO(ksuszyns): SRVCOM-1235 donate those apps to openshift-knative
-        regular: 'quay.io/cardil/knative-serving-showcase:2-send-event',
-        updated: 'quay.io/cardil/knative-serving-showcase-js'
-      }
-    }
-
-    url() {
-      return cy.get('a.co-external-link')
-        .last()
-        .scrollIntoView()
-        .should('have.attr', 'href')
-        .and('include', 'showcase')
-    }
-
-    makeRequest(baseUrl) {
-      const req = {
-        method: 'OPTIONS',
-        url: baseUrl,
-        retryOnStatusCodeFailure: true,
-        failOnStatusCode: true
-      }
-      cy.request(req).then((response) => {
-        expect(response.status).to.eq(200)
-        expect(response.body).to.have.property('version')
-        expect(JSON.stringify(response.body)).to.include('knative-serving-showcase')
-      })
-    }
-
-    checkScale(scale) {
-      const selector = 'div.pf-topology-container__with-sidebar ' +
-        'div.odc-revision-deployment-list__pod svg tspan'
-      const timeout = Cypress.config().defaultCommandTimeout
-      try {
-        // TODO: Remove the increased timeout when https://issues.redhat.com/browse/ODC-5685 is fixed.
-        Cypress.config('defaultCommandTimeout', 300_000)
-        cy.get(selector)
-          .invoke('text')
-          .should((text) => {
-            expect(text).to.eq(`${scale}`)
-          })
-      } finally {
-        Cypress.config('defaultCommandTimeout', timeout)
-      }
-    }
-
-    deployImage({kind = 'regular', clusterLocal = false} = {}) {
-      showcaseKsvc.counter++
-      cy.visit(`/add/ns/${showcaseKsvc.namespace}`)
-      cy.contains('Knative Channel')
-      cy.contains('Event Source')
-      cy.visit(`/deploy-image/ns/${showcaseKsvc.namespace}`)
-      cy.get('input[name=searchTerm]')
-        .type(showcaseKsvc.image[kind])
-      cy.contains('Validated')
-      cy.get('input#form-radiobutton-resources-knative-field').check()
-      cy.get('input#form-checkbox-route-create-field').check()
-      cy.get('input#form-input-application-name-field')
-        .clear()
-        .type(showcaseKsvc.app)
-      cy.get('input#form-input-name-field')
-        .clear()
-        .type(showcaseKsvc.name)
-      if (clusterLocal) {
-        cy.get('input#form-checkbox-route-create-field')
-          .uncheck()
-      }
-      cy.get('button[type=submit]').click()
-      cy.url().should('include', `/topology/ns/${showcaseKsvc.namespace}`)
-      cy.visit(`/topology/ns/${showcaseKsvc.namespace}/list`)
-      cy.get('div.pf-topology-content').contains(showcaseKsvc.name).click()
-      // Make sure the app is running before proceeding.
-      cy.contains('Running')
-    }
-
-    isServiceDeployed() {
-      return new Cypress.Promise((resolve, reject) => {
-        cy.exec(`kubectl get kservice ${showcaseKsvc.name} \
-          -n ${showcaseKsvc.namespace}`, { failOnNonZeroExit: false }).then(result => {
-          resolve(result.code === 0)
-        })
-      })
-    }
-
-    removeApp() {
-      showcaseKsvc.isServiceDeployed().then(deployed => {
-        if (deployed) {
-          showcaseKsvc.doRemoveApp()
-        } else {
-          cy.log("Service isn't deployed, skipping removal.")
-        }
-      })
-    }
-
-    doRemoveApp() {
-      cy.visit(`/topology/ns/${showcaseKsvc.namespace}/list`)
-      cy.get('div.pf-topology-content')
-        .contains(showcaseKsvc.app).click()
-      cy.contains('Actions').click()
-      cy.contains('Delete Application')
-        .should('not.have.class', 'pf-m-disabled')
-        .click()
-      cy.get('input#form-input-resourceName-field')
-        .type(showcaseKsvc.app)
-      cy.get('button#confirm-action.pf-c-button.pf-m-danger').click()
-      cy.contains('No resources found')
-    }
-
-    showServiceDetails() {
-      cy.visit(`/topology/ns/${showcaseKsvc.namespace}/list`)
-      cy.get('div.pf-topology-content')
-        .contains(showcaseKsvc.name).click()
-      cy.contains('Location:')
-        .scrollIntoView()
-    }
-  }
-
+  const environment = new Environment()
+  const openshiftConsole = new OpenshiftConsole()
   const showcaseKsvc = new ShowcaseKservice()
 
   beforeEach(() => {
-    describe('with authenticated via Web Console', cy.login)
-    describe('remove kservice', showcaseKsvc.removeApp)
+    describe('with authenticated via Web Console', () => {
+      openshiftConsole.login()
+    })
+    describe('remove kservice', () => {
+      showcaseKsvc.removeApp()
+    })
   })
 
   afterEach(() => {
-    describe('remove kservice', showcaseKsvc.removeApp)
+    describe('remove kservice', () => {
+      showcaseKsvc.removeApp()
+    })
   })
 
   it('can deploy kservice and scale it', () => {
-    describe('deploy kservice from image', showcaseKsvc.deployImage)
+    describe('deploy kservice from image', () => {
+      showcaseKsvc.deployImage()
+    })
     describe('check automatic scaling of kservice', () => {
       showcaseKsvc.showServiceDetails()
       showcaseKsvc.url().then((url) => {
@@ -154,9 +46,11 @@ describe('OCP UI for Serverless', () => {
   })
 
   it('can route traffic to multiple revisions', () => {
-    describe('deploy kservice from image', showcaseKsvc.deployImage)
+    describe('deploy kservice from image', () => {
+      showcaseKsvc.deployImage()
+    })
     describe('add two revisions to traffic distribution', () => {
-      cy.visit(`/topology/ns/${showcaseKsvc.namespace}/list`)
+      cy.visit(showcaseKsvc.topologyUrl())
       cy.get('div.pf-topology-content')
         .contains(showcaseKsvc.name).click()
       cy.contains('Actions').click()
@@ -168,7 +62,7 @@ describe('OCP UI for Serverless', () => {
       cy.get('button[type=submit]').click()
       cy.url().should('include', showcaseKsvc.namespace)
       cy.contains(showcaseKsvc.app)
-      cy.visit(`/topology/ns/${showcaseKsvc.namespace}/list`)
+      cy.visit(showcaseKsvc.topologyUrl())
       cy.get('div.pf-topology-content')
         .contains(showcaseKsvc.name).click()
       cy.contains('Set traffic distribution', {matchCase: false}).click()
@@ -201,10 +95,9 @@ describe('OCP UI for Serverless', () => {
   })
 
   it('can deploy a cluster-local service', () => {
-    cy.ocpVersion().then((version) => {
-      const range = '>=4.8 || ~4.7.18 || ~4.6.39'
-      cy.onlyOn(version.satisfies(range))
-    })
+    const range = '>=4.8 || ~4.7.18 || ~4.6.39'
+    cy.onlyOn(environment.ocpVersion().satisfies(range))
+
     describe('deploy kservice from image', () => {
       showcaseKsvc.deployImage({clusterLocal: true})
     })
