--- test/ui/cypress/support/ocpversion.js
@@ -1,3 +1,0 @@
-Cypress.Commands.add('ocpVersion', () => {
-  return cy.semver(Cypress.env('OCP_VERSION'))
-})
